% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{body_condition_measurement_estimates}
\alias{body_condition_measurement_estimates}
\title{Sample MCMC output}
\format{
A list with 5 elements:
\describe{
  \item{altimeters}{Posterior samples and summaries for altimeters}
  \item{images}{Posterior samples and summaries for images}
  \item{pixel_error}{Posterior samples and summaries for pixel error 
    component of measurement error model}
  \item{objects}{Posterior samples and summaries for unknown object lengths 
    that were estimated}
 \item{summaries}{\code{data.frame}s with posterior summaries, collated from
    all other list elements.}
}
}
\usage{
body_condition_measurement_estimates
}
\description{
Posterior estimates for lengths and widths of a whale.  See 
\code{help("body_condition")} for computation details.
}
\keyword{datasets}
