% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakFun.R
\name{breakFun}
\alias{breakFun}
\title{Break function (required in models)}
\usage{
breakFun(B, delta)
}
\arguments{
\item{B}{argument to evaluate function at}

\item{delta}{breakpoint location}
}
\description{
Implements Heaviside step function for use in nimble models, H(B) = 1 if 
B <= delta
}
\examples{
breakFun(B = 1, delta = 0)

}
