% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibration}
\alias{calibration}
\title{Calibration data for photogrammetric measurement error model}
\format{
A data frame with 657 rows and 10 variables:
\describe{
  \item{CO.ID}{Training object name}
  \item{Lpix}{Measured length of object in image (in pixels)}
  \item{CO.L}{True length of calibration object (in meters)}
  \item{image}{Name of image that measurements are taken from}
  \item{Baro_Alt}{The altitude (in meters) at which the image was 
        taken, according to the onboard barometer}
  \item{Laser_Alt}{The altitude (in meters) at which the image was 
        taken, according to the attached laser altimeter}
  \item{Focal_Length}{The focal length of the lens when the image was taken, 
        in mm}
  \item{Iw}{The width of the image, in pixels}
  \item{Sw}{The width of the camera sensor, in mm}
  \item{uas}{An identifier for the unoccupied aerial system (UAS) the image 
        and measurements are from}
}
}
\usage{
calibration
}
\description{
A dataset containing photogrammetric measurements of known-length objects, 
used to estimate the characteristics of measurement errors attributed to the
drone platform.
}
\keyword{datasets}
