% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibration2}
\alias{calibration2}
\title{Calibration data for photogrammetric measurement error model}
\format{
A data frame with 46 rows and 9 variables:
\describe{
  \item{RRR.pix}{Measured length of object in image (in pixels)}
  \item{L_train}{True length of calibration object (in meters)}
  \item{Images}{Name of image that measurements are taken from}
  \item{Baro...Ht}{The altitude (in meters) at which the image was 
        taken, according to the onboard barometer}
  \item{Laser_Alt}{The altitude (in meters) at which the image was 
        taken, according to the attached laser altimeter}
  \item{Focal.length}{The focal length of the lens when the image was taken, 
        in mm}
  \item{Iw}{The width of the image, in pixels}
  \item{Sw}{The width of the camera sensor, in mm}
  \item{Aircraft}{An identifier for the unoccupied aerial system (UAS) the 
        image and measurements are from}
}
}
\usage{
calibration2
}
\description{
A dataset containing photogrammetric measurements of known-length objects, 
used to estimate the characteristics of measurement errors attributed to the
drone platform.
}
\keyword{datasets}
