% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sampler.R
\name{calibration_sampler}
\alias{calibration_sampler}
\title{Build an MCMC sampler for calibration data}
\usage{
calibration_sampler(data, priors)
}
\value{
outputs a function to run a sampler, the function arguments are: 
\describe{
 \item{niter}{set the number of iterations}
 \item{burn}{set the number samples to discard}
 \item{thin}{set the thinning rate}
 }
}
\description{
Build an MCMC sampler that only uses calibration data to estimate measurement 
error parameters
}
\examples{
sampler = calibration_sampler(
  data = combine_observations(calibration_data, whale_data),
  priors = list(
    image_altitude = c(min = 0.1, max = 130),
    altimeter_bias = rbind(
      data.frame(altimeter = 'Barometer', mean = 0, sd = 1e2),
      data.frame(altimeter = 'Laser', mean = 0, sd = 1e2)
    ),
    altimeter_variance = rbind(
      data.frame(altimeter = 'Barometer', shape = .01, rate = .01),
      data.frame(altimeter = 'Laser', shape = .01, rate = .01)
    ),
    pixel_variance = c(shape = .01, rate = .01)
  )
)

output = sampler(niter = 1e4)
}
