% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summaries.R
\name{extract_summaries}
\alias{extract_summaries}
\title{Convenience function to combine summary objects from other formatting 
functions}
\usage{
extract_summaries(model_output)
}
\arguments{
\item{model_output}{collection of MCMC output previously formatted using 
functions such as \code{format_altimeter_output}, 
\code{format_image_output}, etc.}
}
\description{
Convenience function to combine summary objects from other formatting 
functions
}
