% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_growth_curve_output.R
\name{format_growth_curve_output}
\alias{format_growth_curve_output}
\title{Extract, summarize, and format growth curve components of model output}
\usage{
format_growth_curve_output(pkg, samples, post_inds)
}
\arguments{
\item{pkg}{analysis package}

\item{samples}{matrix of nimble posterior samples}

\item{post_inds}{array of indices from \code{samples} to use for summarizing
posterior distributions}
}
\description{
Extract, summarize, and format growth curve components of model output
}
