% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_object_output.R
\name{format_object_output}
\alias{format_object_output}
\title{Extract, summarize, and format object size components of model output}
\usage{
format_object_output(pkg, samples, post_inds, prediction_objects)
}
\arguments{
\item{pkg}{analysis package}

\item{samples}{matrix of nimble posterior samples}

\item{post_inds}{array of indices from \code{samples} to use for summarizing
posterior distributions}

\item{prediction_objects}{\code{data.frame} specifying which objects in the 
model had lengths to be estimated}
}
\description{
Extract, summarize, and format object size components of model output
}
