% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curve_sampler.R
\name{growth_curve_sampler}
\alias{growth_curve_sampler}
\title{Build an MCMC sampler that uses calibration data to estimate independent,
unknown lengths}
\usage{
growth_curve_sampler(data, priors)
}
\description{
Build an MCMC sampler that uses calibration data to estimate independent,
unknown lengths
}
