% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_length_sampler.R
\name{independent_length_sampler}
\alias{independent_length_sampler}
\title{Build an MCMC sampler that uses calibration data to estimate independent,
unknown lengths}
\usage{
independent_length_sampler(data, priors)
}
\description{
Build an MCMC sampler that uses calibration data to estimate independent,
unknown lengths
}
