% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_pixel_counts}
\alias{validate_pixel_counts}
\title{Validation checks for pixel data}
\usage{
validate_pixel_counts(x, error = "stop", verbose = TRUE)
}
\description{
Validation checks for pixel data
}
