% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_prediction_objects}
\alias{validate_prediction_objects}
\title{Validation checks for information about object lengths to estimate}
\usage{
validate_prediction_objects(x, error = "stop", verbose = TRUE)
}
\description{
Validation checks for information about object lengths to estimate
}
