% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_training_objects}
\alias{validate_training_objects}
\title{Validation checks for information about known object lengths}
\usage{
validate_training_objects(x, error = "stop", verbose = TRUE)
}
\description{
Validation checks for information about known object lengths
}
