% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whale_info}
\alias{whale_info}
\title{Metadata about whale observations}
\format{
A data frame with 293 rows and 5 variables:
\describe{
  \item{Year}{The year the observations were made}
  \item{Subject}{Unique identifier for the whale}
  \item{Group}{A categorical variable describing the whale's sex}
  \item{ObservedAge}{The whale's age at time of observation}
  \item{AgeType}{Additional details about the ObservedAge measurement}
}
}
\usage{
whale_info
}
\description{
A dataset containing additional information about whales for which 
measurements were taken
}
\keyword{datasets}
