% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whales}
\alias{whales}
\title{Photogrammetric measurements of whales}
\format{
A data frame with 76 rows and 8 variables:
\describe{
  \item{whale_ID}{Unique identifier for the whale, matching 
    \code{whale_info$Subject}
   }
  \item{sex}{The whale's sex}
  \item{Age}{The whale's age at time of observation}
  \item{AgeType}{Additional details about the Age measurement}
  \item{year}{The year in which measurements were made}
  \item{date}{YYYY-MM-DD date string in which measurements were made}
  \item{Image}{Name of image measurements are taken from}
  \item{AltitudeBarometer}{The altitude (in meters) at which the image was 
        taken, according to the onboard barometer}
  \item{AltitudeLaser}{The altitude (in meters) at which the image was 
        taken, according to the attached laser altimeter}
  \item{FocalLength}{The focal length of the lens when the image was taken, 
        in mm}
  \item{ImageWidth}{The width of the image, in pixels}
  \item{SensorWidth}{The width of the camera sensor, in mm}
  \item{UAS}{An identifier for the unoccupied aerial system (UAS) the image 
        and measurements are from}
  \item{TL.pix}{Total length of whale in image (in pixels)}
}
}
\usage{
whales
}
\description{
A dataset containing Photogrammetric measurements of whales.
}
\keyword{datasets}
