% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakFun.R
\name{breakFun}
\alias{breakFun}
\title{Break function (required in models)}
\usage{
breakFun(B, delta)
}
\arguments{
\item{B}{argument to evaluate function at}

\item{delta}{breakpoint location}
}
\value{
1 if \code{B <= delta}, and 0 otherwise
}
\description{
Implements Heaviside step function for use in nimble models, H(B) = 1 if 
B <= delta. For internal use only.  Not intended to be called directly by users.
}
\examples{
breakFun(B = 1, delta = 0) 

}
