% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibration}
\alias{calibration}
\title{Calibration data for photogrammetric measurement error model}
\format{
A data frame with 616 rows and 9 variables:
\describe{
  \item{object_id}{Name of training object}
  \item{image}{Name of image measurements are taken from}
  \item{calibration_measurement}{The length of the object in the image, in 
        pixels}
  \item{true_length}{The known length of the object, in meters}
  \item{barometer_altitude}{The altitude (in meters) at which the image was 
        taken, according to the onboard barometer}
  \item{laser_altitude}{The altitude (in meters) at which the image was 
        taken, according to the attached laser altimeter}
  \item{focal_length}{The focal length of the lens when the image was taken, 
        in mm}
  \item{image_width}{The width of the image, in pixels}
  \item{sensor_width}{The width of the camera sensor, in mm}
}
}
\source{
\url{TODO: fill in data citation}
}
\usage{
calibration
}
\description{
A dataset containing photogrammetric measurements of known-length objects, 
used to estimate the characteristics of measurement errors attributed to the
drone platform.
}
\details{
TODO: update column descriptions
}
\keyword{datasets}
