% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sampler.R
\name{calibration_sampler}
\alias{calibration_sampler}
\title{Build an MCMC sampler that only uses calibration data to estimate measurement 
error parameters}
\usage{
calibration_sampler(data, priors)
}
\description{
Build an MCMC sampler that only uses calibration data to estimate measurement 
error parameters
}
