% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_data.R
\name{flatten_data}
\alias{flatten_data}
\title{Reformat photogrammetric data for model-based analysis}
\usage{
flatten_data(
  data = NULL,
  priors,
  pixel_counts = data$pixel_counts,
  training_objects = data$training_objects,
  image_info = data$image_info
)
}
\arguments{
\item{priors}{\code{list} with elements \code{altitude}, \code{lengths}, 
\code{bias}, and \code{sigma} that parameterize the prior distributions for 
the Bayesian model}

\item{pixel_counts}{\code{data.frame} with columns \code{Subject}, 
\code{Measurement}, \code{Image}, and \code{PixelCount} that describe the 
length measurements taken from images}

\item{training_objects}{\code{data.frame} with columns \code{Subject},
\code{Measurement}, and \code{Length} that describe the known lengths of 
the objects used to calibrate the photogrammetric model}

\item{image_info}{\code{data.frame} with columns \code{Image}, 
\code{AltitudeBarometer}, \code{AltitudeLaser}, \code{FocalLength}, 
\code{ImageWidth}, and \code{SensorWidth} that describe the images used in 
the photogrammetric study}
}
\description{
Assemble \code{data.frame} objects into a format that can be analyzed using 
numerical methods.  This function is analagous to \code{stats::model.matrix},
which generates design matrices for models that are specified via formulas.
}
