% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_length_sampler.R
\name{independent_length_sampler}
\alias{independent_length_sampler}
\title{MCMC sampler for measurements of individuals with no replicates or age information.}
\usage{
independent_length_sampler(data, priors)
}
\arguments{
\item{data}{Photogrammetric data formatted for Xcertainty models, required to
be an object with class \code{obs.parsed}, which can be obtained by running
\code{parse_observations()}}

\item{priors}{\code{list} with components that define the model's prior 
distribution.  See \code{help(flatten_data)} for more details.}
}
\value{
outputs a function to run a sampler, the function arguments are: 
\describe{
 \item{niter}{set the number of iterations}
 \item{burn}{set the number samples to discard}
 \item{thin}{set the thinning rate}
 }
}
\description{
Build an MCMC sampler that uses calibration data to estimate independent,
unknown measurements. This sampler is well suited for when individuals have 
no replicate measurements across time points and have no age information.  It 
can also be used to estimate lengths when there are replicate measurements.
However, the model assumes all Subject/Measurement/Timepoint combinations are
independent, so no strength will be borrowed across temporal replication of a 
subject's measurements, for example.
}
