% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nondecreasing_length_sampler.R
\name{nondecreasing_length_sampler}
\alias{nondecreasing_length_sampler}
\title{Build an MCMC sampler that uses calibration data to estimate independent,
unknown lengths}
\usage{
nondecreasing_length_sampler(data, priors)
}
\description{
Build an MCMC sampler that uses calibration data to estimate independent,
unknown lengths
}
