% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whales}
\alias{whales}
\title{Gray whale metadata}
\format{
A data frame with 826 rows and 14 columns:
\describe{
  \item{whale_ID}{unique individual}
  \item{sex}{Female, Male, or NA}
  \item{Age}{age in years}
  \item{AgeType}{either 'known age' if individual was seen as a calf, or 'min age' from the date of date sighting}
  \item{year}{Year}
  \item{date}{Date}
  \item{Image}{image name}
  \item{AltitudeBarometer}{the barometer altitude adjusted for the launch height of the drone}
  \item{AltitudeLaser}{the altitude recorded by the laser (LiDAR) altimeter}
  \item{FocalLength}{focal length of the camera (mm)}
  \item{ImageWidth}{image width (px)} 
  \item{SensorWidth}{sensor width (mm)}
  \item{UAS}{the unoccupied aircraft system (UAS), or drone, used in data collection}
  \item{TL.pix}{the total body length measurement in pixels}
}
}
\source{
<https://doi.org/10.1111/gcb.17366>
}
\usage{
whales
}
\description{
Gray whale information and metadata that pairs with 'whales' data by "Subject"
}
\keyword{datasets}
